Move file fixtures to tests/fixtures/files.py
